Create QuestionIdentifier
This code was originally designed to search a given text for question words (i.e., who, what, when, etc.). Then add any sentences with those question words to a DataFrame that organized these "questions" by the question word they contain. I would like to refractor this code so that it can use some of the data from this DataFrame to train itself to find questions. Then, I'd like to see how accurate it is at finding new questions in a similar text.